create filtertab component
